+++ Makefile
@@ -31,17 +31,17 @@ SRC		=	src/ft_atoi.c			src/ft_bzero.c		\
 			src/ft_strnstr.c		src/ft_strrchr.c	\
 			src/ft_strtrim.c		src/ft_substr.c		\
 			src/ft_toupper.c		src/ft_tolower.c	\
-												\
+														\
 		src/ft_lstnew.c			src/ft_lstadd_front.c	\
  		src/ft_lstsize.c		src/ft_lstadd_back.c  	\
  		src/ft_lstlast.c		src/ft_lstdelone.c		\
  		src/ft_lstclear.c		src/ft_lstiter.c		\
  		src/ft_lstmap.c									\
 														\
-		src/ft_printf.c			src/puts.c				\
-		src/write.c										\
-		src/get_next_line.c								\
-		src/get_next_line_utils.c						\
+		src/printf/ft_printf.c	src/printf/puts.c		\
+		src/printf/write.c								\
+		src/get_next_line/get_next_line.c				\
+		src/get_next_line/get_next_line_utils.c			\
 
 OBJ		= 	$(SRC:.c=.o)
