+++ src/ft_isprint.c
@@ -14,7 +14,7 @@
 
 int	ft_isprint(int c)
 {
-	if (c >= 32 && c <= 126)
-		return (1);
-	return (0);
+	if (c >= 32 && c <= '~')
+		return (true);
+	return (false);
 }
