+++ src/printf/puts.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "../libft.h"
 
 void	pf_putbase(unsigned long n, char *base, int *result)
 {
